+++ Elements/src/Geometry/Transform.cs
@@ -385,7 +385,27 @@ namespace Elements.Geometry
         {
             var m = new Matrix();
             m.SetupRotate(axis, angle * (Math.PI / 180.0));
-            this.Matrix = this.Matrix * m;
+            this.Matrix *= m;
+        }
+
+        /// <summary>
+        /// Apply a rotation around a to the transform around a point.
+        /// </summary>
+        /// <param name="axis">The axis of rotation.</param>
+        /// <param name="angle">The angle of rotation in degrees.</param>
+        /// <param name="origin">The point around which to rotate.</param>
+        public void Rotate(Vector3 axis, double angle, Vector3 origin)
+        {
+            var mT1 = new Matrix();
+            mT1.SetupTranslation(origin.Negate());
+
+            var mT2 = new Matrix();
+            mT2.SetupTranslation(origin);
+
+            var mR = new Matrix();
+            mR.SetupRotate(axis, angle * (Math.PI / 180.0));
+
+            this.Matrix = this.Matrix * mT1 * mR * mT2;
         }
 
         /// <summary>